+++ packages/eslint-config-nau/rules/errors.js
@@ -9,6 +9,9 @@ module.exports = {
       functions: 'always-multiline',
     }],
 
+    // Enforce “for” loop update clause moving the counter in the right direction.
+    'for-direction': 'error',
+
     // Disallow await inside of loops
     // http://eslint.org/docs/rules/no-await-in-loop
     'no-await-in-loop': 'error',